Change Id type to long and ProfitLoss to float

+++ Models/UserObj.cs
@@ -1,6 +1,6 @@
 public class UserObj
 {
-    public int Id { get; set; }
+    public long Id { get; set; } 
     public string Username { get; set; } = string.Empty;
-    public decimal ProfitLoss { get; set; }
+    public float ProfitLoss { get; set; } 
 }
